clvgl: Add engineLogging true on cppdbg of launch.json
Sometimes an error on the execution of the application may not be
shown on the Debug Console, so it may be interesting to get the full
log by setting this "engineLogging" variable to true. However, it prints
a lot of stuff on Debug Console, so if the user wants a smaller log he
can set it to false.

Signed-off-by: Andre Riesco <andre.riesco@toradex.com>