scripts: init-workspace: Fix hard coded firewall rules
For share the wsl ports during the init, in case we are under WSL,
we need to add the firewall but depending of the values from the
machine. The WSL does not appears to use a fixed IP address, so
we need to get the IP address from the machine and use it to
configure the firewall rules.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>