scripts: PowerShell 2022-2025 ☠️
This project would be nothing without PowerShell, it was the best choice
at the time and I don't regret anything. But now it's time to move on,
move on to something more popular and that has less hate.

Thank you for everything PowerShell, you will always be remembered in
our hearts.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>